Fixing `SidePanelEntryId` broken replacements
This change removes the use of the shadowing file, with a substitution,
as a means to our own values into `SidePanelEntryId`, as this is not
possible anymore with the lates change in the way that enum is
generated, which is involving the use of macros in the same header that
repeat the token in question, causing another unwanted substituion.

Due to this new development, a patch seems to be the only way we can
reasonably add values to this enum now.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/54d8328e82749624a919a7506107c6b331592769

commit 54d8328e82749624a919a7506107c6b331592769
Author: Shibalik Mohapatra <shibalik@chromium.org>
Date:   Tue Oct 10 00:44:46 2023 +0000

    Add a ToString() method to SidePanelEntryKey.

    This method will later be used by SidePanelCoordinator to access the action items dynamically created like in the case of extensions.

    Low-Coverage-Reason:TRIVIAL_CHANGE side_panel_entry_id.h (Tests exist for some of the SIDE_PANEL_ENTRY_IDS) and it is trivial that it should not have to exist for every since entry. The functionality is tested.