[Android] AutofillPopupBridge was removed
We don't need our changes since they were fixing issues in the parent class that are not relevant anymore.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/7f04d239494545dbf9c2ad23cfa71d79cf49bdbe

[Android] Remove AutofillPopupViewAndroid and AutofillPopupBridge.

AutofillPopup is referenced in the following places:
* AwAutofillClient in /android_webview
* AutofillPopupBridge in /chrome/android
* AutofillProvider in /components/android_autofill
* AutofillUnitTest in /chrome/android [test-only]

This CL removes obsolete AutofillPopupViewAndroid implementation
of the AutofillPopupView. Only AutofillKeyboardAccessoryAdapter is
used on Android right now. AutofillPopupBridge is removed as a
dependency of the AutofillPopupViewAndroid.

This is done as part of the effort to move AutofillPopup to
android webview as it's the only place it's being used right now.

Bug: 1448820