Fix breakpad patch for `IsValidCrashReportId`
This patch was introduced in the past to accept a backtrace. Recently
though changes upstream changed exactly the patched lines in question to
use an algorithm-based approach to check for the format of the ID.

This change creates a local function in a shadow file to act as a proxy
to make sure we preserve the `-` check as well as a valid id character.

Original issue: https://github.com/brave/brave-browser/issues/14326

Chromium change:
https://chromium.googlesource.com/chromium/src/+/b999986ff24699b81ecd3f45254a4bdf0332698a

commit b999986ff24699b81ecd3f45254a4bdf0332698a
Author: Peter Kasting <pkasting@chromium.org>
Date:   Wed Oct 25 03:54:27 2023 +0000

    Shorten code by using existing helper APIs, eliminating temps.

    No functional change.

    Bug: none