[CodeHealth] Remove empty classes for permission rules
This change aims to simplify how the permissions are checked in brave
ads. The use of classes derived from a base class, to couple together
disparate functions is less clear to the reader, than the use of free
functions.

An empty class has several drawbacks. The style guide [1] disallow the
use of static function, and member functions into classes which they
have no data relationship with. The use of a horizontal hierarchy
obscures the true relationship these operations have to each other,
which is none.

It is also important to notice that the excessive number of files that
was being used to fragment this class is bound to have some small effect
on build times. There would also likely be binary bloat as well, if the
compiler failed to devirtualise the base class.

Lastly, this change is also removes the use of `base::expected` as a
return type, as that alone was making the call sites for these functions
even more obscure, while any errors were merely being sent for logging
at a single place.

[1] https://google.github.io/styleguide/cppguide.html#Nonmember,_Static_Member,_and_Global_Functions